Create document_changes/revisions w/ subscriptions
This update reworks a bunch of the older paradigms. We can now have
a separate table for each and every document change. This allows us
to better subscribe, apply, paginate and identify individual changes.
In other words, we can now figure out how updated an end users draft is
and apply the changes needed from that point, and easily be able to find
those changes.

Other new cool stuff - using pg-mem for unit testing. It's experimental
but so is this project!